fix: can't use client hook in server
